==> code_pre/06_analyze_predictions.Rmd <==
---
title: "03_analyze_predictions"
output: html_document
---

```{r}
library(tidyverse)
library(pROC)
library(caret)
library(here)
library(blandr)
source(here("code/00_assessment_helper_functions.R"))
```

# Loading files

```{r}
ms_status <- read.csv(here("data/csvs/cavs_ms_status.csv")) %>% 
  mutate(site = as.numeric(str_sub(subject, 2, 2)))
demo <- read.csv(here("data/csvs/cavs_demographics.csv"))

cv_ids <- read.csv(here("cv_results/cv_df.csv")) %>% 
  mutate(cv_index = cv_index + 1)
# competitor_preds <- lapply(list.files(here("cv_results/competitors"),
#              full.names = TRUE), read.csv) %>% 
#   do.call(rbind, .)
competitor_preds <- read.csv(here("cv_results/competitors.csv"))

missing_lesions <- read.csv(here("data/csvs/missing_lesions.csv"))
missing_lesions <- missing_lesions %>% 
  mutate(total_lesions = ifelse(contains_lesions == 0, NA, total_lesions),
         total_cvss = ifelse(contains_cvs == 0, NA, total_cvss))
```

```{r}
cv_preds <- lapply(1:10, function(i) {
  concatenate_pred_csvs(here("cv_results/dependency_1014"), i, cv_ids)
})
train_preds <- lapply(cv_preds, function(cv) cv$train)
test_preds <- lapply(cv_preds, function(cv) cv$test)

total_lesions_are_equal <- sapply(1:length(train_preds), function(i) {
  check_total_lesion_number(train_preds[[i]],
                            test_preds[[i]], 
                            competitor_preds)
})

if (!all(total_lesions_are_equal)) {
  stop("Total lesion numbers are not equivalent for at least one CV split")
}

train_preds_1 <- lapply(train_preds, function(preds) {
  preprocess_preds(preds, competitor_preds)
})
test_preds_1 <- lapply(test_preds, function(preds) {
  preprocess_preds(preds, competitor_preds)
})
```

```{r}
cv_preds <- lapply(1:10, function(i) {
  concatenate_pred_csvs(here("cv_output_conv"), i, cv_ids)
})

train_preds <- lapply(cv_preds, function(cv) cv$train)
test_preds <- lapply(cv_preds, function(cv) cv$test)

total_lesions_are_equal <- sapply(1:length(train_preds), function(i) {
  check_total_lesion_number(train_preds[[i]],
                            test_preds[[i]], 
                            competitor_preds)
})

if (!all(total_lesions_are_equal)) {
  stop("Total lesion numbers are not equivalent for at least one CV split")
}

train_preds_2 <- lapply(train_preds, function(preds) {
  preprocess_preds(preds, competitor_preds)
})
test_preds_2 <- lapply(test_preds, function(preds) {
  preprocess_preds(preds, competitor_preds)
})
```

# Get test ROCs
```{r}
test_all_1 <- do.call(rbind, test_preds_1)
test_all_roc_1 <- get_train_roc(test_all_1)

test_all_2 <- do.call(rbind, test_preds_2)
test_all_roc_2 <- get_train_roc(test_all_2)

layout <- layout(matrix(1:6, 2, 3, byrow = T))
layout <- plot(test_all_roc_1[[13]], print.auc = TRUE)
layout <- plot(test_all_roc_1[[14]], add = TRUE)
layout <- plot(test_all_roc_1[[15]], print.auc = TRUE)
layout <- plot(test_all_roc_1[[16]], add = TRUE)
layout <- plot(test_all_roc_1[[17]], print.auc = TRUE)
layout <- plot(test_all_roc_1[[18]], add = TRUE)
layout <- plot(test_all_roc_2[[13]], print.auc = TRUE)
layout <- plot(test_all_roc_2[[14]], add = TRUE)
layout <- plot(test_all_roc_2[[15]], print.auc = TRUE)
layout <- plot(test_all_roc_2[[16]], add = TRUE)
layout <- plot(test_all_roc_2[[17]], print.auc = TRUE)
layout <- plot(test_all_roc_2[[18]], add = TRUE)

with(test_all_roc_1, roc.test(les_roc, mimosa_roc))
with(test_all_roc_1, roc.test(prl_roc, aprl_roc))
with(test_all_roc_1, roc.test(cvs_roc, acvs_roc))
```

```{r}
train_all_1 <- do.call(rbind, train_preds_1)
train_all_roc_1 <- get_train_roc(train_all_1)

train_all_2 <- do.call(rbind, train_preds_2)
train_all_roc_2 <- get_train_roc(train_all_2)
```


```{r}
no_nas <- filter(test_all_1, !is.na(lesion_true)) %>% 
  mutate(both_pred = 10 * acvs_pred + cvs_pred)
ggplot(no_nas,
       aes(lesion_pred, cvs_pred, color = as.factor(lesion_true + cvs_true))) + 
  geom_point() +
  geom_smooth(method = "lm")

ggplot(no_nas) + 
  geom_density(aes(lesion_pred, fill = as.factor(lesion_true)), alpha = 0.5)
ggplot(no_nas) + 
  geom_density(aes(cvs_pred, fill = as.factor(cvs_true)), alpha = 0.5)
ggplot(no_nas) + 
  geom_density(aes(prl_pred, fill = as.factor(prl_true)), alpha = 0.5)
table(no_nas$cvs_pred > 0.411766, no_nas$acvs_pred != 0)

ggplot(no_nas %>% filter(lesion_true == 1)) + 
  geom_jitter(aes(cvs_pred, acvs_pred, color = as.factor(cvs_true)), height = 0.1)

plot(roc(cvs_true ~ both_pred, data = no_nas))
plot(roc(cvs_true ~ acvs_pred, data = no_nas), col = 2, add = TRUE)
plot(roc(cvs_true ~ cvs_pred, data = no_nas), col = 3, add = TRUE)

```


<!-- # DeLong's test on test ROCs vs competitor ROC -->
<!-- ```{r} -->
<!-- test_all_roc[c(2, 8, 4, 10, 6, 12)] -->

<!-- get_spec_at_sens <- function(roc_obj, sensitivity) { -->
<!--   last_index <- sum(roc_obj$sensitivities > sensitivity) -->
<!--   return(roc_obj$specificities[last_index]) -->
<!-- } -->

<!-- get_spec_at_sens(test_all_roc$prl_roc, 0.8) -->
<!-- get_spec_at_sens(test_all_roc$aprl_roc, 0.8) -->
<!-- ``` -->

<!-- # Get ROCs from raw predictions -->
<!-- ```{r} -->
<!-- train_roc_list <- lapply(train_preds, get_train_roc) -->

<!-- train_roc_acc <- sapply(train_roc_list, extract_roc_acc,  -->
<!--                         simplify = T) %>% t() %>% as.data.frame() %>%  -->
<!--   pivot_longer(cols = everything(), values_to = "train") -->

<!-- test_roc_list <- mlapply(get_test_roc, test_preds, train_roc_list) -->

<!-- test_roc_acc <- sapply(test_roc_list, extract_roc_acc,  -->
<!--                        simplify = T) %>% t() %>% as.data.frame() %>%  -->
<!--   pivot_longer(cols = everything(), values_to = "test") -->

<!-- roc_acc_df <- cbind(train_roc_acc,  -->
<!--                     test_roc_acc,  -->
<!--                     cv_id = rep(1:length(train_preds),  -->
<!--                                 each = 15))[, -1] %>%  -->
<!--   rename(metric = name) %>%  -->
<!--   pivot_longer(cols = c("train", "test")) -->
<!-- ``` -->

<!-- # Correct predictions where lesion prediction and PRL/CVS prediction differ -->
<!-- ```{r} -->
<!-- mapply(get_disparate_preds, train_preds, train_roc_list) %>% signif(digits = 3) -->

<!-- train_preds_corr <- mlapply(change_disparate_preds, -->
<!--                             train_preds, train_roc_list) -->
<!-- test_preds_corr <- mlapply(change_disparate_preds, -->
<!--                            test_preds, train_roc_list) -->

<!-- train_roc_list_corr <- lapply(train_preds_corr, get_train_roc) -->

<!-- train_roc_acc_corr <- sapply(train_roc_list_corr, extract_roc_acc, -->
<!--                              simplify = T) %>% t() %>% as.data.frame() %>% -->
<!--   pivot_longer(cols = everything(), values_to = "train") -->

<!-- test_roc_list_corr <- mlapply(get_test_roc, test_preds_corr, train_roc_list_corr) -->

<!-- test_roc_acc_corr <- sapply(test_roc_list_corr, extract_roc_acc, -->
<!--                             simplify = T) %>% t() %>% as.data.frame() %>% -->
<!--   pivot_longer(cols = everything(), values_to = "test") -->

<!-- roc_acc_df_corr <- cbind(train_roc_acc_corr, -->
<!--                          test_roc_acc_corr, -->
<!--                          cv_id = rep(1:length(train_preds_corr), -->
<!--                                      each = 15))[, -1] %>% -->
<!--   rename(metric = name) %>% -->
<!--   pivot_longer(cols = c("train", "test")) -->
<!-- ``` -->

# ROC by site
```{r}
test_all_site <- test_all_1 %>% mutate(site = as.numeric(str_sub(as.character(subject), 2, 2)))
by_subject <- test_all_site %>% 
  group_by(subject) %>% 
  summarize(n_cand = n(),
            n_lesion = sum(lesion_true, na.rm = T),
            n_prl = sum(prl_true),
            n_cvs = sum(cvs_true, na.rm = T)) 
for (i in 0:9) {
  if (i == 1) {
    plot(roc(cvs_true ~ cvs_pred, data = test_all_site %>% filter(site == i)))
  }
  print(i)
  plot(roc(cvs_true ~ cvs_pred, data = test_all_site %>% filter(site == i)), add = TRUE, col = i)
}
plot(roc(cvs_true ~ cvs_pred, data = test_all_site %>% filter(site == 9)), add = TRUE, col = 2)
```

# Get training thresholds
```{r}
# Get thresholds from comprehensive training set (Training set from all 10 CVs) 
train_preds_all <- do.call(rbind, train_preds_1)
all_roc <- get_train_roc(train_preds_all)

les_thresh <- get_threshold(all_roc$les_roc, 3)
prl_thresh <- get_threshold(all_roc$prl_roc, 3)
cvs_thresh <- get_threshold(all_roc$cvs_roc, 3)

mimosa_thresh <- get_threshold(all_roc$mimosa_roc, 3)
aprl_thresh <- get_threshold(all_roc$aprl_roc, 3)
acvs_thresh <- get_threshold(all_roc$acvs_roc, 3)
```

```{r}
test_predictions <- test_all_1 %>% 
  mutate(lesion_yn = lesion_pred >= les_thresh[1], 
         prl_yn = prl_pred >= prl_thresh[1], 
         cvs_yn = cvs_pred >= cvs_thresh[1], 
         mimosa_yn = mimosa_pred >= mimosa_thresh[1], 
         aprl_yn = aprl_pred >= aprl_thresh[1], 
         acvs_yn = acvs_pred >= acvs_thresh[1])
write.csv(test_predictions, here("cv_results/test_predictions.csv"))
```

```{r}
test_all_joined <- test_all_1 %>% 
  group_by(subject) %>% 
  summarise(n_lesions = sum(lesion_pred > les_thresh[1]),
            n_prls = sum(prl_pred > prl_thresh[2]), 
            n_cvss = sum(cvs_pred > cvs_thresh[2]),
            n_lesion_true = sum(lesion_true),
            n_prl_true = sum(prl_true),
            n_cvs_true = sum(cvs_true),
            prob_lesions = sum(lesion_pred),
            prob_prls = sum(prl_pred),
            prob_cvss = sum(cvs_pred),
            mean_lesions = mean(lesion_pred),
            mean_prls = mean(prl_pred),
            mean_cvss = mean(cvs_pred)) %>% 
  left_join(ms_status, by = "subject") %>%
  left_join(missing_lesions, by = "subject") %>% 
  left_join(demo, by = "subject") %>% 
  mutate(ms_diagnosis = ifelse(ms_diagnosis == "Yes", 1, 0))
test_all_joined$ms_diagnosis <- as.factor(test_all_joined$ms_diagnosis)

ggplot(test_all_joined) + 
  geom_jitter(aes(n_lesions, total_lesions.x, color = ms_diagnosis)) +
  geom_abline(intercept = 0, slope = 1)
ggplot(test_all_joined) + 
  geom_jitter(aes(n_prls, total_prls, color = ms_diagnosis)) +
  geom_abline(intercept = 0, slope = 1)
ggplot(test_all_joined) + 
  geom_jitter(aes(n_cvss, total_cvss, color = ms_diagnosis)) +
  geom_abline(intercept = 0, slope = 1)

ggplot(test_all_joined) + 
  geom_jitter(aes(prob_lesions, total_lesions, color = ms_diagnosis)) +
  geom_abline(intercept = 0, slope = 1)
ggplot(test_all_joined) + 
  geom_jitter(aes(prob_prls, total_prls, color = ms_diagnosis)) +
  geom_abline(intercept = 0, slope = 1)
ggplot(test_all_joined) + 
  geom_jitter(aes(prob_cvss, total_cvss, color = ms_diagnosis)) +
  geom_abline(intercept = 0, slope = 1)

ggplot(test_all_joined) + 
  geom_jitter(aes(mean_lesions, total_lesions, color = ms_diagnosis))
ggplot(test_all_joined) + 
  geom_jitter(aes(mean_prls, total_prls, color = ms_diagnosis))
ggplot(test_all_joined) + 
  geom_jitter(aes(mean_cvss, total_cvss, color = ms_diagnosis))

ggplot(test_all_joined) +
  geom_boxplot(aes(ms_diagnosis, n_lesions))
ggplot(test_all_joined) +
  geom_boxplot(aes(ms_diagnosis, n_prls))
ggplot(test_all_joined) +
  geom_boxplot(aes(ms_diagnosis, n_cvss))

ggplot(test_all_joined) +
  geom_boxplot(aes(ms_diagnosis, prob_lesions))
ggplot(test_all_joined) +
  geom_boxplot(aes(ms_diagnosis, prob_prls))
ggplot(test_all_joined) +
  geom_boxplot(aes(ms_diagnosis, prob_cvss))

ggplot(test_all_joined) +
  geom_boxplot(aes(ms_diagnosis, mean_lesions))
ggplot(test_all_joined) +
  geom_boxplot(aes(ms_diagnosis, mean_prls))
ggplot(test_all_joined) +
  geom_boxplot(aes(ms_diagnosis, mean_cvss))

with(test_all_joined, cor.test(n_lesions, total_lesions.y, method = "spearman"))
with(test_all_joined, cor.test(n_prls, total_prls.y, method = "spearman"))
with(test_all_joined, cor.test(n_cvss, total_cvss.y, method = "spearman"))

with(test_all_joined, cor.test(prob_lesions, total_lesions.y, method = "spearman"))
with(test_all_joined, cor.test(prob_prls, total_prls.y, method = "spearman"))
with(test_all_joined, cor.test(prob_cvss, total_cvss.y, method = "spearman"))

with(test_all_joined, cor.test(mean_lesions, total_lesions, method = "spearman"))
with(test_all_joined, cor.test(mean_prls, total_prls, method = "spearman"))
with(test_all_joined, cor.test(mean_cvss, total_cvss, method = "spearman"))

glm(ms_diagnosis ~ age_yrs_calc + sex_female + total_cvss.y + total_prls.y + total_lesions.y, 
    data = test_all_joined, family = binomial()) %>% summary()
glm(ms_diagnosis ~ age_yrs_calc + sex_female + total_cvss, 
    data = test_all_joined, family = binomial()) %>% summary()
glm(ms_diagnosis ~ age_yrs_calc + sex_female + total_lesions, 
    data = test_all_joined, family = binomial()) %>% summary()
glm(ms_diagnosis ~ age_yrs_calc + sex_female + total_prls, 
    data = test_all_joined, family = binomial()) %>% summary()
glm(ms_diagnosis ~ age_yrs_calc + sex_female + n_cvss + n_prls + n_lesions, 
    data = test_all_joined, family = binomial()) %>% summary()
glm(ms_diagnosis ~ age_yrs_calc + sex_female + n_cvss, 
    data = test_all_joined, family = binomial()) %>% summary()
glm(ms_diagnosis ~ age_yrs_calc + sex_female + n_prls, 
    data = test_all_joined, family = binomial()) %>% summary()
glm(ms_diagnosis ~ age_yrs_calc + sex_female + n_lesions, 
    data = test_all_joined, family = binomial()) %>% summary()
glm(ms_diagnosis ~ age_yrs_calc + sex_female + prob_cvss + prob_prls + prob_lesions, 
    data = test_all_joined, family = binomial()) %>% summary()
glm(ms_diagnosis ~ log(prob_cvss), 
    data = test_all_joined, family = binomial()) %>% summary()
glm(ms_diagnosis ~ age_yrs_calc + sex_female + prob_prls, 
    data = test_all_joined, family = binomial()) %>% summary()
glm(ms_diagnosis ~ age_yrs_calc + sex_female + prob_lesions, 
    data = test_all_joined, family = binomial()) %>% summary()
```

# Plot patch
```{r}
example_subject <- list.files(list.files(here("data/processed_05"), 
                                         full.names = TRUE)[2], 
                              full.names = TRUE)
plot_patch <- function(path_list, corner) {
  if (length(path_list) == 1)
    path_list <- list(path_list)
  lapply(path_list, function(path) {
    browser()
    image_array <- as.array(check_ants(path))
    patch <- image_array[corner[1]:(corner[1] + 23), corner[2]:(corner[2] + 23), corner[3]]
    par(mar=c(0,0,0,0))
    image(patch, col = grey(seq(0, 1, length = 256)), axes = FALSE)
    par(mar=c(0,0,0,0))
    patch <- image_array[corner[1], corner[2]:(corner[2] + 23), corner[3]:(corner[3] + 23)]
    image(patch, col = grey(seq(0, 1, length = 256)), axes = FALSE)
    par(mar=c(0,0,0,0))
    patch <- image_array[corner[1]:(corner[1] + 23), corner[2], corner[3]:(corner[3] + 23)]
    image(patch, col = grey(seq(0, 1, length = 256)), axes = FALSE)
  })
}

plot_patch(example_subject[c(18, 8, 6, 12, 17, 15)], c(65, 112, 230))

plot_patch(example_subject[15], c(65, 112, 230))
```
